Add watchbootstrap.sh and watchcluster.sh
This change adds two shell scripts useful for watching a v4 cluster
while it is bootstrapping. The first, watchbootstrap.sh, will invoke an
indefiniate loop where it will automatically determine the IP address of
the bootstrap host based on the terraform state file and then ssh into
that host and display the journal logs for the bootkube service. The
second will use watch to continually display the output of multiple oc
commands: the cluster version and status, any unusual events, the list of
nodes, the list of pending csr requests, and existing cluster operators
that are not in the desired state.