Use ephemeral ssh keys instead of openshift-dev.pem
This commit removes the reliance on a deprecated openshift key pair.
For each cluster deployment, a new key pair will be created and added to
AWS EC2.  To ensure ease of use, a new command, `ssh`,  was added to
`test-cluster.sh`.  This command can be used to ssh to any of the nodes
in the cluster deployed by Ansible.